pyproject.toml
[project]
name = "BISTMon"
version = "0.2.0"
description = "Analyze state of MCU-Boards by utilizing GPIOs to characterize external circuitry."
keywords = ["built-in self-test", "BIST", "MCU", "MCU-Boards", "Monitor", "Framework", "generalized"]

authors = [
    {name = "Päpke, Tim Carlo",email = "tim.paepke@stud.tu-darmstadt.de"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "pyserial (>=3.5,<4.0)",
    "cbor2 (>=5.7.0,<6.0.0)",
    "xxhash (>=3.5.0,<4.0.0)",
    "crc32c (>=2.7.1,<3.0.0)",
    "crcmod (>=1.7,<2.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "chromalog",
    "typer",
    "defusedxml",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-click",
    "coverage",
]

dev = [
    "testsoftware-monitor[test]",
    "pre-commit",
    "pre-commit-uv", # check with `pre-commit --version`
    "pandas-stubs",  # for pyright with pandas
]

[project.scripts]
bistmon = "bistmon.cli:cli"

[build-system]
requires = ["setuptools>=61.2", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
platforms = ["unix", "linux", "osx", "cygwin", "win32", "win64"]
package-dir = {"bistmon" = "./src/"}
